requirements_test.txt
# Test requirements for AegisBot integration
pytest-homeassistant-custom-component
pytest-asyncio
pytest-cov
async_timeout
aiohttp
